move the controller under internal namespace
